+++ src/kingdom/kingdom.ts
@@ -303,7 +303,7 @@ class KingdomApp extends FormApplication<FormApplicationOptions & KingdomOptions
         Hooks.on('canvasReady', this.sceneChange.bind(this));
         Hooks.on('createToken', this.sceneChange.bind(this));
         Hooks.on('deleteToken', this.sceneChange.bind(this));
-        document.addEventListener('kmAppliedModifierFromChat', this.sceneChange.bind(this));
+        document.addEventListener('kmReRenderKingdomSheet', this.sceneChange.bind(this));
         const $html = html[0];
         $html.querySelectorAll('.km-nav a')?.forEach(el => {
             el.addEventListener('click', (event) => {
@@ -702,7 +702,7 @@ class KingdomApp extends FormApplication<FormApplicationOptions & KingdomOptions
         Hooks.off('canvasReady', this.sceneChange);
         Hooks.off('createToken', this.sceneChange);
         Hooks.off('deleteToken', this.sceneChange);
-        document.removeEventListener('kmAppliedModifierFromChat', this.sceneChange);
+        document.removeEventListener('kmReRenderKingdomSheet', this.sceneChange);
         return super.close(options);
     }
 
@@ -884,7 +884,6 @@ class KingdomApp extends FormApplication<FormApplicationOptions & KingdomOptions
 
     private async saveKingdom(kingdom: Partial<Kingdom>): Promise<void> {
         await saveKingdom(this.sheetActor, kingdom);
-        this.render();
     }
 
     private getKingdom(): Kingdom {